GPT as module
