+++ examples/gpt_enum.rs
@@ -1,54 +1,10 @@
 /// Illustrates #[xdevs::model_enum]: a GPT model where the processor is
 /// chosen at build time between a fast and a slow variant, without any
 /// conditional logic in the coupled model.
-use xdevs::simulation::Simulable;
-use xdevs::AbstractSimulator;
-
-mod generator {
-    pub struct Generator {
-        sigma: f64,
-        period: f64,
-        count: usize,
-    }
-
-    impl xdevs::Component for Generator {
-        type Kind = xdevs::AtomicKind;
-        type Input = xdevs::Port<bool, 1>;
-        type Output = xdevs::Port<usize, 1>;
-    }
-
-    impl xdevs::Atomic for Generator {
-        fn delta_int(&mut self) {
-            self.count += 1;
-            self.sigma = self.period;
-            println!("[G] generated job {}", self.count);
-        }
-        fn lambda(&self, output: &mut Self::Output) {
-            output.add_value(self.count).unwrap();
-        }
-        fn ta(&self) -> f64 {
-            self.sigma
-        }
-        fn delta_ext(&mut self, elapsed: f64, input: &Self::Input) {
-            self.sigma -= elapsed;
-            if let Some(&stop) = input.get_values().last() {
-                if stop {
-                    self.sigma = f64::INFINITY;
-                }
-            }
-        }
-    }
-
-    impl Generator {
-        pub fn new(period: f64) -> Self {
-            Self {
-                sigma: 0.0,
-                period,
-                count: 0,
-            }
-        }
-    }
-}
+use xdevs::{
+    gpt::{Generator, Transducer},
+    simulation::{AbstractSimulator, Simulable},
+};
 
 mod processor {
     pub struct FastProcessor {
@@ -135,7 +91,7 @@ mod processor {
                 if self.job.is_none() {
                     println!("[P-slow] received job {}", job);
                     self.job = Some(job);
-                    self.sigma = self.time * 2.0; // Slow processor takes twice as long
+                    self.sigma = self.time * 2.0;
                 }
             }
         }
@@ -158,74 +114,11 @@ mod processor {
     }
 }
 
-mod transducer {
-    #[derive(xdevs::Bag)]
-    pub struct TransducerInput {
-        pub in_generator: xdevs::Port<usize, 1>,
-        pub in_processor: xdevs::Port<usize, 1>,
-    }
-
-    pub struct Transducer {
-        sigma: f64,
-        clock: f64,
-        n_generated: usize,
-        n_processed: usize,
-    }
-
-    impl xdevs::Component for Transducer {
-        type Kind = xdevs::AtomicKind;
-        type Input = TransducerInput;
-        type Output = xdevs::Port<bool, 1>;
-    }
-
-    impl xdevs::Atomic for Transducer {
-        fn delta_int(&mut self) {
-            self.clock += self.sigma;
-            let (acceptance, throughput) = if self.n_processed > 0 {
-                (
-                    self.n_processed as f64 / self.n_generated as f64,
-                    self.n_processed as f64 / self.clock,
-                )
-            } else {
-                (0.0, 0.0)
-            };
-            println!(
-                "[T] acceptance: {:.2}, throughput: {:.2} jobs/s",
-                acceptance, throughput
-            );
-            self.sigma = f64::INFINITY;
-        }
-        fn lambda(&self, output: &mut Self::Output) {
-            output.add_value(true).unwrap();
-        }
-        fn ta(&self) -> f64 {
-            self.sigma
-        }
-        fn delta_ext(&mut self, elapsed: f64, input: &Self::Input) {
-            self.sigma -= elapsed;
-            self.clock += elapsed;
-            self.n_generated += input.in_generator.get_values().len();
-            self.n_processed += input.in_processor.get_values().len();
-        }
-    }
-
-    impl Transducer {
-        pub fn new(obs_time: f64) -> Self {
-            Self {
-                sigma: obs_time,
-                clock: 0.0,
-                n_generated: 0,
-                n_processed: 0,
-            }
-        }
-    }
-}
-
 #[xdevs::coupled]
 pub struct GPT {
-    generator: generator::Generator,
+    generator: Generator,
     processor: processor::Processor,
-    transducer: transducer::Transducer,
+    transducer: Transducer,
 }
 
 impl xdevs::Component for GPT {
@@ -251,29 +144,25 @@ impl xdevs::Coupled for GPT {
 }
 
 fn run_gpt(processor: processor::Processor) {
-    let period = 1.;
-    let obs_time = 10.;
+    const PERIOD: f64 = 1.;
+    const OBS_TIME: f64 = 10.;
 
     let label = match &processor {
         processor::Processor::Fast(_) => "fast",
         processor::Processor::Slow(_) => "slow",
     };
     println!("\n--- GPT with {} processor ---", label);
-    let gpt = GPT::build(
-        generator::Generator::new(period),
-        processor,
-        transducer::Transducer::new(obs_time),
-    );
+    let gpt = GPT::build(Generator::new(PERIOD), processor, Transducer::new(OBS_TIME));
     let mut simulator = gpt.to_simulator();
     let config = xdevs::simulation::Config::new(0.0, 14.0, 1.0, None);
     simulator.simulate_rt(&config, xdevs::simulation::std::sleep(&config), |_| {});
 }
 
 fn main() {
-    let proc_time = 1.1;
-    let fast = processor::Processor::Fast(processor::FastProcessor::new(proc_time).to_simulator());
+    const PROC_TIME: f64 = 1.1;
+    let fast = processor::Processor::Fast(processor::FastProcessor::new(PROC_TIME).to_simulator());
     run_gpt(fast);
 
-    let slow = processor::Processor::Slow(processor::SlowProcessor::new(proc_time).to_simulator());
+    let slow = processor::Processor::Slow(processor::SlowProcessor::new(PROC_TIME).to_simulator());
     run_gpt(slow);
 }
