+++ src/gpt.rs
@@ -1,0 +1,277 @@
+/// Generator that produces jobs at a fixed period until told to stop.
+pub struct Generator {
+    sigma: f64,
+    period: f64,
+    count: usize,
+}
+
+impl xdevs::Component for Generator {
+    type Kind = xdevs::AtomicKind;
+    type Input = xdevs::Port<bool, 1>;
+    type Output = xdevs::Port<usize, 1>;
+}
+
+impl xdevs::Atomic for Generator {
+    fn delta_int(&mut self) {
+        self.count += 1;
+        self.sigma = self.period;
+    }
+
+    fn lambda(&self, output: &mut Self::Output) {
+        #[cfg(feature = "std")]
+        std::println!("[G] sending job {}", self.count);
+        output.add_value(self.count).unwrap();
+    }
+
+    fn ta(&self) -> f64 {
+        self.sigma
+    }
+
+    fn delta_ext(&mut self, elapsed: f64, input: &Self::Input) {
+        self.sigma -= elapsed;
+        if let Some(&stop) = input.get_values().last() {
+            #[cfg(feature = "std")]
+            std::println!("[G] received stop: {}", stop);
+            if stop {
+                self.sigma = f64::INFINITY;
+            }
+        }
+    }
+}
+
+impl Generator {
+    pub fn new(period: f64) -> Self {
+        Self {
+            sigma: 0.0,
+            period,
+            count: 0,
+        }
+    }
+}
+
+/// Processor that receives a job, processes it for a fixed duration, then outputs it.
+pub struct Processor {
+    sigma: f64,
+    time: f64,
+    job: Option<usize>,
+}
+
+impl xdevs::Component for Processor {
+    type Kind = xdevs::AtomicKind;
+    type Input = xdevs::Port<usize, 1>;
+    type Output = xdevs::Port<usize, 1>;
+}
+
+impl xdevs::Atomic for Processor {
+    fn delta_int(&mut self) {
+        self.sigma = f64::INFINITY;
+        if self.job.is_some() {
+            #[cfg(feature = "std")]
+            std::println!("[P] processed job {}", self.job.unwrap());
+            self.job = None;
+        }
+    }
+
+    fn lambda(&self, output: &mut Self::Output) {
+        if let Some(job) = self.job {
+            output.add_value(job).unwrap();
+        }
+    }
+
+    fn ta(&self) -> f64 {
+        self.sigma
+    }
+
+    fn delta_ext(&mut self, elapsed: f64, input: &Self::Input) {
+        self.sigma -= elapsed;
+        if let Some(&job) = input.get_values().last() {
+            #[cfg(feature = "std")]
+            std::print!("[P] received job {}", job);
+            if self.job.is_none() {
+                #[cfg(feature = "std")]
+                std::println!(" (idle)");
+                self.job = Some(job);
+                self.sigma = self.time;
+            } else {
+                #[cfg(feature = "std")]
+                std::println!(" (busy)");
+            }
+        }
+    }
+}
+
+impl Processor {
+    pub fn new(time: f64) -> Self {
+        Self {
+            sigma: 0.0,
+            time,
+            job: None,
+        }
+    }
+}
+
+/// Input bag for the Transducer.
+#[derive(xdevs::Bag)]
+pub struct TransducerInput {
+    pub in_generator: xdevs::Port<usize, 1>,
+    pub in_processor: xdevs::Port<usize, 1>,
+}
+
+/// Transducer that observes generated and processed jobs, computes metrics,
+/// and sends a stop signal to the Generator.
+pub struct Transducer {
+    sigma: f64,
+    clock: f64,
+    n_generated: usize,
+    n_processed: usize,
+}
+
+impl xdevs::Component for Transducer {
+    type Kind = xdevs::AtomicKind;
+    type Input = TransducerInput;
+    type Output = xdevs::Port<bool, 1>;
+}
+
+impl xdevs::Atomic for Transducer {
+    fn delta_int(&mut self) {
+        self.clock += self.sigma;
+        let (acceptance, throughput) = if self.n_processed > 0 {
+            (
+                self.n_processed as f64 / self.n_generated as f64,
+                self.n_processed as f64 / self.clock,
+            )
+        } else {
+            (0.0, 0.0)
+        };
+        #[cfg(feature = "std")]
+        std::println!(
+            "[T] acceptance: {:.2}, throughput: {:.2}",
+            acceptance,
+            throughput
+        );
+        self.sigma = f64::INFINITY;
+    }
+
+    fn lambda(&self, output: &mut Self::Output) {
+        output.add_value(true).unwrap();
+    }
+
+    fn ta(&self) -> f64 {
+        self.sigma
+    }
+
+    fn delta_ext(&mut self, elapsed: f64, input: &Self::Input) {
+        self.sigma -= elapsed;
+        self.clock += elapsed;
+        self.n_generated += input.in_generator.get_values().len();
+        self.n_processed += input.in_processor.get_values().len();
+    }
+}
+
+impl Transducer {
+    pub fn new(obs_time: f64) -> Self {
+        Self {
+            sigma: obs_time,
+            clock: 0.0,
+            n_generated: 0,
+            n_processed: 0,
+        }
+    }
+
+    pub fn acceptance(&self) -> f64 {
+        if self.n_generated > 0 {
+            self.n_processed as f64 / self.n_generated as f64
+        } else {
+            0.0
+        }
+    }
+
+    pub fn throughput(&self) -> f64 {
+        if self.clock > 0.0 {
+            self.n_processed as f64 / self.clock
+        } else {
+            0.0
+        }
+    }
+}
+
+#[xdevs::coupled]
+pub struct GPT {
+    generator: Generator,
+    processor: Processor,
+    transducer: Transducer,
+}
+
+impl xdevs::Component for GPT {
+    type Kind = xdevs::CoupledKind;
+    type Input = ();
+    type Output = ();
+}
+
+impl xdevs::Coupled for GPT {
+    fn ic(
+        from: &xdevs::component::coupled::ComponentsOutput<Self>,
+        to: &mut xdevs::component::coupled::ComponentsInput<Self>,
+    ) {
+        from.generator.couple(&mut to.processor).unwrap();
+        from.processor
+            .couple(&mut to.transducer.in_processor)
+            .unwrap();
+        from.generator
+            .couple(&mut to.transducer.in_generator)
+            .unwrap();
+        from.transducer.couple(&mut to.generator).unwrap();
+    }
+}
+
+#[xdevs::coupled]
+pub struct EF {
+    generator: Generator,
+    transducer: Transducer,
+}
+
+impl xdevs::Component for EF {
+    type Kind = xdevs::CoupledKind;
+    type Input = xdevs::Port<usize, 1>;
+    type Output = xdevs::Port<usize, 1>;
+}
+
+impl xdevs::Coupled for EF {
+    fn ic(
+        from: &xdevs::component::coupled::ComponentsOutput<Self>,
+        to: &mut xdevs::component::coupled::ComponentsInput<Self>,
+    ) {
+        from.generator
+            .couple(&mut to.transducer.in_generator)
+            .unwrap();
+        from.transducer.couple(&mut to.generator).unwrap();
+    }
+    fn eic(from: &Self::Input, to: &mut xdevs::component::coupled::ComponentsInput<Self>) {
+        from.couple(&mut to.transducer.in_processor).unwrap();
+    }
+    fn eoc(from: &xdevs::component::coupled::ComponentsOutput<Self>, to: &mut Self::Output) {
+        from.generator.couple(to).unwrap();
+    }
+}
+
+#[xdevs::coupled]
+pub struct EFP {
+    ef: EF,
+    processor: Processor,
+}
+
+impl xdevs::Component for EFP {
+    type Kind = xdevs::CoupledKind;
+    type Input = ();
+    type Output = ();
+}
+
+impl xdevs::Coupled for EFP {
+    fn ic(
+        from: &xdevs::component::coupled::ComponentsOutput<Self>,
+        to: &mut xdevs::component::coupled::ComponentsInput<Self>,
+    ) {
+        from.ef.couple(&mut to.processor).unwrap();
+        from.processor.couple(&mut to.ef).unwrap();
+    }
+}
